refactor(core): cleanup jsonify code
