+++ src/jsonify/stringify.ts
@@ -1,4 +1,6 @@
+import { encodeBase64 } from "@std/encoding/base64";
 import {
+  HOLE,
   INFINITY_NEG,
   INFINITY_POS,
   NAN,
@@ -6,7 +8,6 @@ import {
   UNDEFINED,
   ZERO_NEG,
 } from "./constants.ts";
-import { HOLE } from "./constants.ts";
 
 export type Stringifiers = Record<
   string,
@@ -109,7 +110,7 @@ function serializeInner(
     } else if (value instanceof RegExp) {
       str += `["RegExp",${JSON.stringify(value.source)}, "${value.flags}"]`;
     } else if (value instanceof Uint8Array) {
-      str += `["Uint8Array","${b64encode(value.buffer)}"]`;
+      str += `["Uint8Array","${encodeBase64(value)}"]`;
     } else if (value instanceof Set) {
       const items = new Array(value.size);
       let i = 0;
@@ -146,44 +147,4 @@ function serializeInner(
 
   out[idx] = str;
   return idx;
-}
-
-// deno-fmt-ignore
-const base64abc = [
-  "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O",
-  "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", "a", "b", "c", "d",
-  "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p", "q", "r", "s",
-  "t", "u", "v", "w", "x", "y", "z", "0", "1", "2", "3", "4", "5", "6", "7",
-  "8", "9", "+", "/",
-];
-
-/**
- * CREDIT: https://gist.github.com/enepomnyaschih/72c423f727d395eeaa09697058238727
- * Encodes a given Uint8Array, ArrayBuffer or string into RFC4648 base64 representation
- */
-export function b64encode(buffer: ArrayBufferLike): string {
-  const uint8 = new Uint8Array(buffer);
-  let result = "",
-    i;
-  const l = uint8.length;
-  for (i = 2; i < l; i += 3) {
-    result += base64abc[uint8[i - 2] >> 2];
-    result += base64abc[((uint8[i - 2] & 0x03) << 4) | (uint8[i - 1] >> 4)];
-    result += base64abc[((uint8[i - 1] & 0x0f) << 2) | (uint8[i] >> 6)];
-    result += base64abc[uint8[i] & 0x3f];
-  }
-  if (i === l + 1) {
-    // 1 octet yet to write
-    result += base64abc[uint8[i - 2] >> 2];
-    result += base64abc[(uint8[i - 2] & 0x03) << 4];
-    result += "==";
-  }
-  if (i === l) {
-    // 2 octets yet to write
-    result += base64abc[uint8[i - 2] >> 2];
-    result += base64abc[((uint8[i - 2] & 0x03) << 4) | (uint8[i - 1] >> 4)];
-    result += base64abc[(uint8[i - 1] & 0x0f) << 2];
-    result += "=";
-  }
-  return result;
 }
